requirements.txt
streamlit>=1.40.0
yfinance>=0.2.40
plotly>=5.20.0
pandas>=2.1.0
scipy>=1.11.0
statsmodels>=0.14.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
